+++ models/pendulum.py
@@ -21,7 +21,7 @@ class Pendulum(BaseEnvironment, gym.Env):
         self.variable_names = ['position', 'velocity']
         self.plot_dim = [0, 1]
 
-        self.max_torque = np.array([1])
+        self.max_torque = np.array([1], dtype=np.float32)
 
         # Pendulum parameters
         self.delta = 0.05
